it works

+++ common/src/main/java/org/whispersystems/curve25519/java/sc_cmov.java
@@ -8,12 +8,12 @@ public class sc_cmov {
     Preconditions: b in {0,1}.
     */
 
-    public static void sc_cmov(byte[] f, byte[] g, byte b) {
+    public static void sc_cmov(byte[] f, byte[] g, int b) {
         int count = 32;
         byte[] x = new byte[32];
         for (count = 0; count < 32; count++)
             x[count] = (byte) (f[count] ^ g[count]);
-        b = (byte) (-b);
+        b = -b;
         for (count = 0; count < 32; count++)
             x[count] &= b;
         for (count = 0; count < 32; count++)
